Фронтенд разбит на отдельные angular-модели. Мелкие правки по бэкенду.
