+++ src/main/java/ru/vyazankin/market/configs/SecurityConfig.java
@@ -24,7 +24,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         http
                 .csrf().disable()
                 .authorizeRequests()
-                .antMatchers("/api/v1/**").authenticated()
+                .antMatchers("/api/v1/products/**").permitAll()
+                //.antMatchers("/api/v1/**").authenticated()
                 .antMatchers("/openapi/v1/**").permitAll()
                 .antMatchers("/h2-console/**").permitAll()
                 .anyRequest().permitAll()
