+++ src/main/java/ru/vyazankin/market/dto/CartDto.java
@@ -21,5 +21,7 @@ public class CartDto {
                 .collect(Collectors.toList());
 
         totalPrice = cart.getTotalCartPrice();
+
+
     }
 }
