+++ src/main/resources/static/cart/cart.html
@@ -1,0 +1,45 @@
+<h3>Корзина</h3>
+
+<table class="table table-hover" cellpadding="0" cellspacing="0">
+    <thead class="thead-light">
+    <tr>
+        <td>Название</td>
+        <td>Цена за единицу</td>
+        <td>Количество</td>
+        <td>Стоимость</td>
+    </tr>
+    </thead>
+    <tbody>
+    <tr ng-repeat="ci in cart.cartItemDtos">
+        <td>{{ci.title}}</td>
+        <td>{{ci.price}}</td>
+        <td width="150px">
+            <div class="input-group">
+                <button class="btn btn-primary" type="button" ng-click="cartRemoveProduct(ci.productId)">-</button>
+                <input type="text" class="form-control" placeholder="0" aria-label="" value="{{ci.quantity}}">
+                <button class="btn btn-primary" type="button" ng-click="cartAddProduct(ci.productId)">+</button>
+            </div>
+        </td>
+        <td>{{ci.itemsPrice}}</td>
+    </tr>
+    <tr>
+        <td></td>
+        <td></td>
+        <td></td>
+        <td>{{cart.totalPrice}}</td>
+    </tr>
+    </tbody>
+</table>
+
+
+
+<div ng-show="!isCartEmpty()" class="form-group">
+
+    <a class="btn btn-primary" role="button" href="#!/order_confirmation">Оформить заказ</a>
+    <button class="btn-primary btn mr-2" type="button" ng-click="setAllowClearCart()">Очистить корзину</button>
+
+    <div ng-show="allowClearCart == true">
+        Нажмите Да, чтобы подтвердить очистку корзины!
+        <button class="btn-danger btn mr-2" type="button" ng-click="cartClear()">Да</button>
+    </div>
+</div>
