+++ src/main/resources/static/home/home.html
@@ -1,0 +1,3 @@
+<span>
+    Главная страница магазина
+</span>
