+++ src/main/resources/static/home/home.js
@@ -1,0 +1,3 @@
+angular.module('app').controller('homeController', function ($scope, $http, $localStorage) {
+    const contextPath = 'http://localhost:8189/happy';
+});
