+++ src/main/resources/static/index.js
@@ -1,223 +1,96 @@
+(function ($localStorage) {
+    'use strict';
 
-angular.module('app', []).controller('indexController', function ($scope, $http) {
-    const contentPath = 'http://localhost:8189/market/'
-    let appPath = contentPath + 'api/v1';
-    let appOpenApiPath = contentPath + 'openapi/v1';
-    let authPath = contentPath +'auth'
+    angular
+        .module('app', ['ngRoute', 'ngStorage'])
+        .config(config)
+        .run(run);
 
-    $scope.currentPage = 1;
-    $scope.allowClearCart = false;
-    $scope.authPresent = false;
-    $scope.newUserPage = false;
-    $scope.user = {name:'',password:''};
-    $scope.orderDetails = {deliveryAddress: ''}
-    $scope.newUser = {name: '', password:'', email:'', error:''}
-
-    $scope.fillProducts = function () {
-        $http({
-            url: appPath + '/products',
-            method: 'GET',
-            params: {
-                page: $scope.currentPage,
-                products: 5
-            }
-        }).then(function (response) {
-            $scope.ProductsPage = response.data;
-        });
-    };
-
-    $scope.fillCart = function (){
-        $http({
-            url: appPath + '/cart',
-            method: 'GET'
+    function config($routeProvider, $httpProvider) {
+        $routeProvider
+            .when('/', {
+                templateUrl: 'home/home.html',
+                controller: 'homeController'
             })
-        .then(function (response) {
-            $scope.Cart = response.data;
-            $scope.allowClearCart = false;
-        })
+            .when('/products', {
+                templateUrl: 'products/products.html',
+                controller: 'productsController'
+            })
+            .when('/cart', {
+                templateUrl: 'cart/cart.html',
+                controller: 'cartController'
+            })
+            .when('/order_confirmation', {
+                templateUrl: 'order_confirmation/order_confirmation.html',
+                controller: 'orderConfirmationController'
+            })
+            //.when('/user_orders/:orderId', {
+            .when('/user_orders', {
+                templateUrl: 'user_orders/user_orders.html',
+                controller: 'userOrdersController'
+            })
+            .when('/orders', {
+                templateUrl: 'orders/orders.html',
+                controller: 'ordersController'
+            })
+            .otherwise({
+                redirectTo: '/'
+            });
+
     }
 
-    $scope.fillOrders = function (){
-        $http({
-            url: appPath + '/orders',
-            method: 'GET'
-        })
-            .then(function (response) {
-                $scope.orderList = response.data;
-                $scope.allowClearCart = false;
-            })
+    function run($rootScope, $http, $localStorage) {
+        if ($localStorage.currentUser) {
+            $http.defaults.headers.common.Authorization = 'Bearer ' + $localStorage.currentUser.token;
+        }
     }
+})();
 
+angular.module('app').controller('indexController', function ($scope, $http, $location, $localStorage) {
 
-    $scope.submitCreateNewProduct = function () {
-        $http.post(appPath + '/products', $scope.newProduct)
-            .then(function (response) {
-                $scope.newProduct = null;
-                $scope.fillProducts();
+    const contextPath = 'http://localhost:8189/market';
+    $scope.user = {username:'', password:''};
+
+    $scope.tryToAuth = function () {
+        console.log($scope.user);
+        $http.post(contextPath + '/auth', $scope.user)
+            .then(function successCallback(response) {
+                if (response.data.token) {
+                    $http.defaults.headers.common.Authorization = 'Bearer ' + response.data.token;
+                    $localStorage.currentUser = {username: $scope.user.username, token: response.data.token};
+
+                    $scope.currentUserName = $scope.user.username;
+
+                    $scope.user.username = null;
+                    $scope.user.password = null;
+                }
+            }, function errorCallback(response) {
             });
     };
 
-    $scope.deleteProductById = function(id) {
-        $http.delete(appPath + '/products/' + id)
-            .then(function (){
-                $scope.fillProducts();
-            })
-    }
+    $scope.tryToLogout = function () {
+        $scope.clearUser();
+        if ($scope.user === null) return;
 
-    $scope.nextPage = function (){
-        if ($scope.ProductsPage == null) return;
-        if (!$scope.ProductsPage.last) {
-            $scope.currentPage++;
-            $scope.fillProducts();
+        if ($scope.user.username) {
+            $scope.user.username = null;
         }
-    }
-
-    $scope.previousPage = function (){
-        if ($scope.ProductsPage == null) return;
-        if (!$scope.ProductsPage.first) {
-            $scope.currentPage--;
-            $scope.fillProducts();
+        if ($scope.user.password) {
+            $scope.user.password = null;
         }
-    }
+        $location.path('/home/');
+    };
 
-    $scope.cartAddProduct = function (productId){
-        $http(
-            {
-                url: appPath + "/cart/add/" + productId,
-                method: 'GET'
-            })
-            .then(function (){
-                $scope.fillCart();
-            })
-    }
+    $scope.clearUser = function () {
+        delete $localStorage.currentUser;
+        $http.defaults.headers.common.Authorization = '';
+    };
 
-
-    $scope.cartRemoveProduct = function(productId) {
-        $http(
-            {
-                url: appPath + "/cart/remove/" + productId,
-                method: 'GET'
-            })
-            .then(function (){
-                $scope.fillCart();
-            })
-    }
-
-    $scope.setAllowClearCart = function (){
-        $scope.allowClearCart = true;
-    }
-
-    $scope.cartClear = function (){
-        $http(
-            {
-                url: appPath + "/cart/clear",
-                method: 'GET'
-            })
-            .then(function (){
-                $scope.fillCart();
-
-            });
-    }
-
-    $scope.isUserLoggedIn = function (){
-        return $scope.authPresent;
-    }
-
-    $scope.authorizationRequest = function (){
-        $http.post(authPath,$scope.user)
-            .then(
-                function successCallback(response) {
-                    if (response.data.token) {
-                        $http.defaults.headers.common.Authorization = 'Bearer ' + response.data.token;
-                        $scope.user.password = null;
-                        $scope.authorized = true;
-
-                        $scope.fillProducts();
-                        $scope.fillCart();
-                        $scope.fillOrders();
-
-                        $scope.authPresent = true;
-                        $scope.newUserPage = false;
-                    }
-                },
-                function errorCallback(response) {
-                    window.alert("Error");
-                    $scope.authPresent = false;
-                });
-    }
-
-
-    $scope.makeOrder = function (){
-        console.log($scope.orderDetails.deliveryAddress)
-        $http(
-            {
-                url: appPath + '/orders/makeorder',
-                method: 'POST',
-                params: {
-                    deliveryAddress: $scope.orderDetails.deliveryAddress
-                }
-            })
-            .then(
-                function (){
-                    $scope.fillOrders();
-                    $scope.cartClear();
-                }
-            );
-    }
-
-    $scope.showNewUserPage = function (){
-        $scope.newUserPage = true;
-    }
-
-    $scope.hideNewUserPage = function (){
-        $scope.newUserPage = false;
-    }
-
-
-    $scope.newUserRequest = function (){
-
-        if ($scope.newUser === null) {
-            return;
+    $scope.isUserLoggedIn = function () {
+        if ($localStorage.currentUser) {
+            return true;
+        } else {
+            return false;
         }
-
-        var haserr = false;
-        $scope.newUser.error = '';
-
-        if ($scope.newUser.name === null || $scope.newUser.name.trim() === ''){
-            createLoginInput.placeholder = 'Необходимо ввести логин!'
-            haserr = true;
-        }
-        if ($scope.newUser.password === null || $scope.newUser.password.trim() === ''){
-            createPasswordInput.placeholder = 'Необходимо ввести пароль!'
-            haserr = true;
-        }
-        if ($scope.newUser.email === null || $scope.newUser.email.trim() === ''){
-            createEmailInput.placeholder = 'Необходимо ввести адрес электронной почты!'
-            haserr = true;
-        }
-
-
-
-        if (haserr) {
-            $scope.newUser.error = 'При регистрации пользователя произошли ошибки!';
-            return;
-        }
-
-        $http.post(appOpenApiPath + '/users', $scope.newUser)
-            .then(
-                function (response){
-                    if (response.data.error === ''){
-                        $scope.hideNewUserPage();
-                        $scope.user.username = response.data.name;
-                        $scope.user.password = response.data.password;
-                        $scope.newUser.password = '';
-                    } else {
-                        $scope.newUser.error = response.data.error;
-                    }
-
-                }
-            )
-    }
-
-
+    };
 });
