+++ src/main/resources/static/order_confirmation/order_confirmation.html
@@ -1,0 +1,41 @@
+<h3>Оформление заказа</h3>
+
+<table class="table table-hover" cellpadding="0" cellspacing="0">
+    <thead class="thead-light">
+    <tr>
+        <td>Название</td>
+        <td>Цена за единицу</td>
+        <td>Количество</td>
+        <td>Стоимость</td>
+    </tr>
+    </thead>
+    <tbody>
+    <tr ng-repeat="ci in cart.cartItemDtos">
+        <td>{{ci.title}}</td>
+        <td>{{ci.price}}</td>
+        <td width="150px">
+            <div class="input-group">
+                <input type="text" class="form-control" placeholder="0" aria-label="" value="{{ci.quantity}}" readonly>
+            </div>
+        </td>
+        <td>{{ci.itemsPrice}}</td>
+    </tr>
+    <tr>
+        <td></td>
+        <td></td>
+        <td></td>
+        <td>{{cart.totalPrice}}</td>
+    </tr>
+    </tbody>
+</table>
+
+<form ng-submit="makeOrder()">
+    <div class="form-group">
+        <label for="addressInput">Введите адрес доставки</label>
+        <input type="text" class="form-control" id="addressInput" aria-describedby="addressHelp"
+               placeholder="Моксва, Мавзолей, д.1" ng-model="orderDetails.deliveryAddress">
+    </div>
+    <div class="form-group">
+        <button class="btn btn-primary" type="submit">Заказать</button>
+    </div>
+</form>
