+++ src/main/resources/static/products/products.html
@@ -1,0 +1,30 @@
+<h3>Список товаров к покупке</h3>
+
+<table class="table table-hover" cellpadding="0" cellspacing="0">
+    <thead class="thead-light">
+    <tr>
+        <td>Название</td>
+        <td>Цена</td>
+        <td></td>
+    </tr>
+    </thead>
+    <tbody>
+    <tr ng-repeat="p in ProductsPage.content">
+        <td>{{p.title}}</td>
+        <td>{{p.price}}</td>
+        <td>
+            <button ng-if="isUserLoggedIn()" class="btn btn-primary" ng-click="cartAddProduct(p.id)">Добавить в корзину</button>
+        </td>
+    </tr>
+    </tbody>
+</table>
+
+<nav aria-label="Page navigation">
+    <ul class="pagination">
+        <li class="page-item" ng-click="previousPage()"><a class="page-link">Previous</a></li>
+        <li class="page-item" ng-click="showProducts()"><a class="page-link">{{currentPage}}</a></li>
+        <li class="page-item" ng-click="nextPage()"><a class="page-link">Next</a></li>
+    </ul>
+</nav>
+
+<a class="btn btn-primary" role="button" href="#!/cart">Перейти к корзине</a>
