+++ src/main/resources/static/user_orders/user_orders.html
@@ -1,0 +1,19 @@
+<h3>Список ваших заказов</h3>
+<table class="table table-hover" cellpadding="0" cellspacing="0">
+    <thead class="thead-light">
+    <tr>
+        <td>Номер</td>
+        <td>Количество позиций</td>
+        <td>Общая стоимость</td>
+        <td>Адрес доставки</td>
+    </tr>
+    </thead>
+    <tbody>
+    <tr ng-repeat="ord in orderList">
+        <td>{{ord.id}}</td>
+        <td>{{ord.totalItems}}</td>
+        <td>{{ord.totalPrice}}</td>
+        <td>{{ord.deliveryAddress}}</td>
+    </tr>
+    </tbody>
+</table>
